Update: got to 100% coverage for addColl and added a new "Unit Test Invariant" P-48: Adding collateral improves Nominal ICR
Note: the Properties.sol file and the invariant function names are still not very practical 🤔. In particular, sometimes it is better to use assertGt than assertWithMsg, for example, but the invariant functions just return true/false. Also, there's quite some code repetition between the revert messages and the properties descriptions. It might be better to have the invariants return a bytes memory string that can be directly fed into the revert message to help with debugging. I'll deal with this refactoring later